+++ src/app/classes/EventLog/event.ts
@@ -21,7 +21,7 @@ import { jsonObject, jsonMember, jsonArrayMember } from 'typedjson';
 export class Event {
     @jsonArrayMember(EventLogAttribute)
     private _attributes: Array<EventLogAttribute>;
-    @jsonMember
+    @jsonMember(String)
     private _activity: string;
 
     public get attributes(): Array<EventLogAttribute> {
