+++ src/app/classes/EventLog/eventlogattribute.ts
@@ -2,14 +2,14 @@ import 'reflect-metadata';
 import { jsonObject, jsonMember } from 'typedjson';
 
 export abstract class EventLogAttribute {
-    @jsonMember
+    @jsonMember(String)
     key: string = '';
     abstract value: any;
 }
 
 @jsonObject
 export class StringAttribute extends EventLogAttribute {
-    @jsonMember
+    @jsonMember(String)
     value: string;
 
     constructor(value: string, key: string) {
@@ -21,7 +21,7 @@ export class StringAttribute extends EventLogAttribute {
 
 @jsonObject
 export class DateAttribute extends EventLogAttribute {
-    @jsonMember
+    @jsonMember(Date)
     value: Date;
 
     constructor(value: Date, key: string) {
@@ -33,7 +33,7 @@ export class DateAttribute extends EventLogAttribute {
 
 @jsonObject
 export class IntAttribute extends EventLogAttribute {
-    @jsonMember
+    @jsonMember(Number)
     value: number;
 
     constructor(value: number, key: string) {
@@ -45,7 +45,7 @@ export class IntAttribute extends EventLogAttribute {
 
 @jsonObject
 export class FloatAttribute extends EventLogAttribute {
-    @jsonMember
+    @jsonMember(Number)
     value: number;
 
     constructor(value: number, key: string) {
@@ -57,7 +57,7 @@ export class FloatAttribute extends EventLogAttribute {
 
 @jsonObject
 export class BooleanAttribute extends EventLogAttribute {
-    @jsonMember
+    @jsonMember(Boolean)
     value: boolean;
 
     constructor(value: boolean, key: string) {
