+++ src/app/workers/log-parser.worker.ts
@@ -1,10 +1,10 @@
 /// <reference lib="webworker" />
-import { XesParser } from '../classes/parser/xesParser';
 import { TypedJSON } from 'typedjson';
 import { EventLog } from '../classes/EventLog/eventlog';
+import { LogParser } from '../classes/parser/logParser';
 
 onmessage = function (data) {
-    const parser = new XesParser();
+    const parser = new LogParser();
     const result = parser.parse(data.data);
     const serializer = new TypedJSON(EventLog);
     const jsonResult = serializer.stringify(result);
